+++ library/Imbo/EventListener/AccessToken.php
@@ -118,16 +118,20 @@ class AccessToken implements ListenerInterface {
 
         // First the the raw un-encoded URI, then the URI as is
         $uris = array($request->getRawUri(), $request->getUriAsIs());
-        $privateKey = $request->getPrivateKey();
+        $privateKeys = $event->getUserLookup()->getPrivateKeys(
+            $request->getPublicKey()
+        ) ?: [];
 
         foreach ($uris as $uri) {
             // Remove the access token from the query string as it's not used to generate the HMAC
             $uri = rtrim(preg_replace('/(?<=(\?|&))accessToken=[^&]+&?/', '', $uri), '&?');
 
-            $correctToken = hash_hmac('sha256', $uri, $privateKey);
+            foreach ($privateKeys as $privateKey) {
+                $correctToken = hash_hmac('sha256', $uri, $privateKey);
 
-            if ($correctToken === $token) {
-                return;
+                if ($correctToken === $token) {
+                    return;
+                }
             }
         }