+++ library/Imbo/EventListener/Cors.php
@@ -91,7 +91,7 @@ class Cors implements ListenerInterface {
         foreach ($this->params['allowedMethods'] as $resource => $methods) {
             foreach ($methods as $method) {
                 $eventName = $resource . '.' . strtolower($method);
-                $events[$eventName] = array('invoke' => 100);
+                $events[$eventName] = array('invoke' => 1000);
             }
 
             // Always enable the listener for the OPTIONS method
@@ -161,10 +161,21 @@ class Cors implements ListenerInterface {
             $allowedMethods = array_merge($allowedMethods, $this->params['allowedMethods'][$resource]);
         }
 
+        $allowedHeaders = array('Content-Type', 'Accept');
+
+        $requestHeaders = $request->headers->get('Access-Control-Request-Headers', '');
+        $requestHeaders = array_map('trim', explode(',', $requestHeaders));
+
+        foreach ($requestHeaders as $header) {
+            if (strpos($header, 'x-imbo') === 0) {
+                $allowedHeaders[] = implode('-', array_map('ucfirst', explode('-', $header)));;
+            }
+        }
+
         $response->headers->add(array(
             'Access-Control-Allow-Origin' => $origin,
             'Access-Control-Allow-Methods' => implode(', ', $allowedMethods),
-            'Access-Control-Allow-Headers' => 'Content-Type, Accept',
+            'Access-Control-Allow-Headers' => implode(', ', $allowedHeaders),
             'Access-Control-Max-Age' => (int) $this->params['maxAge'],
         ));