+++ tests/behat/imbo-configs/image-content-negotiation-disabled.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * This file is part of the Imbo package
+ *
+ * (c) Christer Edvartsen <cogo@starzinger.net>
+ *
+ * For the full copyright and license information, please view the LICENSE file that was
+ * distributed with this source code.
+ */
+
+/**
+ * Disable content negotiation for images
+ */
+return array(
+    'contentNegotiateImages' => false,
+);